fix dashboard 404 and nodes uptime parse error
Dashboard was hitting /api/system/stats/system-stats which doesn't exist;
correct path is /api/system/stats. Node and dashboard uptime fields changed
from Long to Double in DTOs — the backend returns fractional seconds (e.g.
2338592.49); domain models stay Long via .toLong().

+++ composeApp/src/commonMain/kotlin/art/yniyniyni/freedomwave/data/api/dto/DashboardDto.kt
@@ -13,7 +13,7 @@ data class SystemStatsResponse(
 data class SystemStatsData(
     @SerialName("cpu")         val cpu: CpuData,
     @SerialName("memory")      val memory: MemoryData,
-    @SerialName("uptime")      val uptime: Long,
+    @SerialName("uptime")      val uptime: Double,
     @SerialName("timestamp")   val timestamp: Long,
     @SerialName("users")       val users: UsersData,
     @SerialName("onlineStats") val onlineStats: OnlineStatsData,
